+++ src/Axle.Dto/TerminateConnectionReason.cs
@@ -1,0 +1,11 @@
+﻿// (c) Lykke Corporation 2019 - All rights reserved. No copying, adaptation, decompiling, distribution or any other form of use permitted.
+
+namespace Axle.Dto
+{
+    public enum TerminateConnectionReason
+    {
+        Other,
+        DifferentDevice,
+        DifferentTab
+    }
+}