+++ src/Axle/Hubs/SessionHub.cs
@@ -3,36 +3,32 @@
 namespace Axle.Hubs
 {
     using System;
-    using System.Collections.Generic;
+    using System.Net;
     using System.Threading.Tasks;
     using Axle.Constants;
     using Axle.Contracts;
     using Axle.Dto;
     using Axle.Extensions;
-    using Axle.Persistence;
     using Axle.Services;
     using IdentityModel;
     using Microsoft.AspNetCore.Authentication.JwtBearer;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.SignalR;
+    using PermissionsManagement.Client.Extensions;
     using Serilog;
 
     [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme, Policy = AuthorizationPolicies.AccountOwnerOrSupport)]
     public class SessionHub : Hub
     {
-        private readonly IRepository<string, HubCallerContext> connectionRepository;
-        private readonly ISessionLifecycleService sessionLifecycleService;
-        private readonly IHubContext<SessionHub> sessionHubContext;
+        private readonly IHubConnectionService hubConnectionService;
+        private readonly ISessionService sessionService;
 
         public SessionHub(
-            IRepository<string, HubCallerContext> connectionRepository,
-            ISessionLifecycleService sessionLifecycleService,
-            IHubContext<SessionHub> sessionHubContext)
+            IHubConnectionService hubConnectionService,
+            ISessionService sessionService)
         {
-            this.connectionRepository = connectionRepository;
-            this.sessionLifecycleService = sessionLifecycleService;
-            this.sessionHubContext = sessionHubContext;
-            this.sessionLifecycleService.OnCloseConnections += this.TerminateConnections;
+            this.hubConnectionService = hubConnectionService;
+            this.sessionService = sessionService;
         }
 
         public static string Name => "/session";
@@ -49,33 +45,51 @@ namespace Axle.Hubs
 
             var isSupportUser = this.Context.User.IsSupportUser(accountId);
 
-            await this.sessionLifecycleService.OpenConnection(this.Context.ConnectionId, userName, accountId, clientId, token, isSupportUser);
+            await this.hubConnectionService.OpenConnection(this.Context, userName, accountId, clientId, token, isSupportUser);
 
             Log.Information($"New connection established. User: {userName}, ID: {this.Context.ConnectionId}.");
-            this.connectionRepository.Add(this.Context.ConnectionId, this.Context);
         }
 
         public override Task OnDisconnectedAsync(Exception exception)
         {
-            var sub = this.Context.User.FindFirst(JwtClaimTypes.Subject).Value;
+            var userName = this.Context.User.FindFirst(JwtClaimTypes.Name).Value;
 
-            Log.Information($"Connection closed. User: {sub}, ID: {this.Context.ConnectionId}.");
-            this.connectionRepository.Remove(this.Context.ConnectionId);
-            this.sessionLifecycleService.CloseConnection(this.Context.ConnectionId);
+            Log.Information($"Connection closed. User: {userName}, ID: {this.Context.ConnectionId}.");
+            this.hubConnectionService.CloseConnection(this.Context.ConnectionId);
+
             return base.OnDisconnectedAsync(exception);
         }
 
-        private void TerminateConnections(IEnumerable<string> connections, SessionActivityType reason)
+        public async Task<bool> SignOut()
         {
-            foreach (var connection in connections)
-            {
-                if (reason == SessionActivityType.DifferentDeviceTermination)
-                {
-                    this.sessionHubContext.Clients.Clients(connection)
-                                       .SendAsync("concurrentSessionTermination", StatusCode.IF_ATH_502, StatusCode.IF_ATH_502.ToMessage()).Wait();
-                }
+            var userName = this.Context.User.FindFirst(JwtClaimTypes.Name).Value;
+            var accountId = this.Context.GetHttpContext().Request.Query["account_id"];
+            var isSupportUser = this.Context.User.IsSupportUser(accountId);
 
-                this.connectionRepository.Get(connection).Abort();
+            var response = await this.sessionService.TerminateSession(userName, accountId, isSupportUser, SessionActivityType.SignOut);
+
+            return response.Status == TerminateSessionStatus.Terminated;
+        }
+
+        public Task<OnBehalfChangeResponse> SetOnBehalfAccount(string accountId)
+        {
+            this.ThrowIfUnauthorized(matchAllPermissions: true, Permissions.OnBehalfSelection);
+
+            if (!this.hubConnectionService.TryGetSessionId(this.Context.ConnectionId, out int sessionId))
+            {
+                throw new HubException("The current connection has not been registered");
+            }
+
+            return this.sessionService.UpdateOnBehalfState(sessionId, accountId);
+        }
+
+        private void ThrowIfUnauthorized(bool matchAllPermissions = false, params string[] permissions)
+        {
+            if (!this.Context.User.IsAuthorized(matchAllPermissions, permissions))
+            {
+                throw new HubException(
+                    $"Action Forbidden ({(int)HttpStatusCode.Forbidden}). " +
+                    "The user does not have the required permissions.");
             }
         }
     }